requirements.txt
python-binance
numpy
backtrader
flask
dateparser
pytz
datetime